requirements.txt
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Faker==26.1.0
idna==3.3
iniconfig==1.1.1
jsonschema==4.22.0
packaging==21.3
pluggy==1.4.0
py==1.11.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==8.1.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
pytest-repeat==0.9.1
python-dotenv~=0.21.1
pydantic==2.7.1
kafka-python~=2.0.2
pytest-dependency~=0.5.1
pytest-rerunfailures~=14.0
confluent-kafka==2.3.0
pytest-playwright==0.4.4
allure-pytest>=2.8.18